fix(test): move ProgressBar test to atoms, test calculatepercentage() in ProgressBar
